Added javadoc and change null handling to rely more on Optional

+++ SavingsGoalManager/src/main/java/com/brenner/budgetmanager/savingsgoals/SavingsGoalsApi.java
@@ -3,15 +3,16 @@
  */
 package com.brenner.budgetmanager.savingsgoals;
 
-import java.util.List;
-import java.util.Optional;
-
+import com.brenner.budgetmanager.exception.NotFoundException;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import lombok.extern.slf4j.Slf4j;
+import java.util.List;
+import java.util.Optional;
 
 /**
+ * API to the Savings Goal business service
  *
  * @author dbrenner
  * 
@@ -24,12 +25,38 @@ public class SavingsGoalsApi {
 	@Autowired
 	SavingsGoalsBusinessService service;
 	
+	/**
+	 * Access to the list of savings goals.
+	 *
+	 * @return An unordered list of all savings goals.
+	 */
 	@GetMapping(path = "/savingsgoals")
 	public List<SavingsGoal> getAllSavingsGoals() {
 		List<SavingsGoal> goals = this.service.getAllSavingsGoals();
 		return goals;
 	}
 	
+	/**
+	 * Access to retrieve the specific default goal. A 404 is generated if there is no default goal.
+	 *
+	 * @return The goal marked as a default.
+	 */
+	@GetMapping(path="/savingsgoals/defaultgoal")
+	public SavingsGoal getDefaultGoal() {
+		Optional<SavingsGoal> optGoal = this.service.findDefaultGoal();
+		if (optGoal.isEmpty()) {
+			throw new NotFoundException("Default goal does not exist.");
+		}
+		
+		return optGoal.get();
+	}
+	
+	/**
+	 * Access to save a new goal
+	 *
+	 * @param savingsGoal The goal data to persist
+	 * @return The persisted object including unique identifier
+	 */
 	@PostMapping(path="/savingsgoals")
 	public SavingsGoal addSavingsGoal(@RequestBody SavingsGoal savingsGoal) {
 		if (savingsGoal.getCurrentBalance() == null) {
@@ -39,12 +66,34 @@ public class SavingsGoalsApi {
 		return newGoal;
 	}
 	
+	/**
+	 * Access to allocate amounts to goals associated with a specific deposit.
+	 *
+	 * @param savingsGoalDepositAllocations The list of allocations
+	 */
+	@PutMapping(path="/savingsgoals/allocateDeposit")
+	public void allocateDepositToGoals(@RequestBody List<SavingsGoalDepositAllocation> savingsGoalDepositAllocations) {
+		this.service.allocateDepositToGoals(savingsGoalDepositAllocations);
+	}
+	
+	/**
+	 * Access to update a goal. If the goal to update doesn't exist a 404 will be generated.
+	 *
+	 * @param id The goal's unique identifier
+	 * @param savingsGoal The goal data to persist
+	 * @return The goal after persistence
+	 */
 	@PutMapping(path="/savingsgoals/{id}")
 	public SavingsGoal updateSavingsGoal(@PathVariable Integer id, @RequestBody SavingsGoal savingsGoal) {
 		
 		log.debug("Call to update SavingsGoal: " + savingsGoal);
 		
-		SavingsGoal sg = this.service.getSavingsGoalById(id);
+		Optional<SavingsGoal> optionalSavingsGoal = this.service.getSavingsGoalById(id);
+		if (optionalSavingsGoal.isEmpty()) {
+			throw new NotFoundException("Savings goal with id " + id + " does not exist.");
+		}
+		
+		SavingsGoal sg = optionalSavingsGoal.get();
 		sg.setCurrentBalance(savingsGoal.getCurrentBalance());
 		sg.setGoalName(savingsGoal.getGoalName());
 		sg.setInitialBalance(savingsGoal.getInitialBalance());
@@ -54,13 +103,20 @@ public class SavingsGoalsApi {
 		return this.service.updateSavingsGoal(sg);
 	}
 	
+	/**
+	 * Access to delete a specific goal. A 404 will be generated if the goal is not found.
+	 *
+	 * @param id The goal's unique identifier.
+	 */
 	@DeleteMapping(path="/savingsgoals/{id}")
 	public void deleteSavingsGoal(@PathVariable Integer id) {
 		
-		SavingsGoal savingsGoal = this.service.getSavingsGoalById(id);
-		if (savingsGoal == null) {
-			///
+		Optional<SavingsGoal> optionalSavingsGoal = this.service.getSavingsGoalById(id);
+		if (optionalSavingsGoal.isEmpty()){
+			throw new NotFoundException("Savings goal with id " + id + " does not exist.");
 		}
+		
+		SavingsGoal savingsGoal = optionalSavingsGoal.get();
 		this.service.deleteSavingsGoal(id);
 	}
